Added signup and login page with Supabase Auth
Supabase has inbuilt Authentication.
I used that so now people can use the signup page to create an account (the password is hashed and stored securely by supabase, we can't read it)

They can then also login, the credentials are checked.

There's the option to send out confirmation emails and all that, however we need a different email provider as the supabase one is rate-limited and really only for testing